DeleteSymbolDialog - Split into partials
